requirements.txt
python-telegram-bot
python-dotenv
slack-bolt
slack-sdk
whatsapp-api-client-python
flask